requirements.txt
fastapi==0.111.1
uvicorn[standard]==0.23.2
psycopg2-binary==2.9.9
python-dotenv==1.0.0   # optional, if you use .env for DB credentials
